requirements.txt
asgiref==3.6.0
Django==4.1.5
django-ckeditor==6.5.1
django-filter==22.1
django-js-asset==2.0.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3